id 타입 Long으로 통일

+++ src/main/java/com/ll/gramgram/boundedContext/member/repository/MemberRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 // JpaRepository<Member, Long> => Member == member 테이블, Long == member 의 주키의 타입
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByUsername(String username);
-    Optional<Member> findByInstaMemberId(Integer userId);
+    Optional<Member> findByInstaMemberId(Long userId);
 }
